--- a/src/h.js
+++ b/src/h.js
@@ -1,10 +1,12 @@
 function addChild(children, val) {
+  if (val == null || typeof val === 'boolean') return
+
   if (Array.isArray(val)) {
     for (var j = 0; j < val.length; j++) {
       children.push(val[j])
     }
   }
-  else if (val != null && val !== true && val !== false) {
+  else {
     children.push(typeof val === 'number' ? val + '' : val)
   }
 }
@@ -12,14 +14,16 @@ function addChild(children, val) {
 export default function(tag, data, values) {
   var children = []
 
-  if (Array.isArray(values)) {
-    for (var i = 0; i < values.length; i++) {
-      addChild(children, values[i])
+  if (arguments.length > 2) {
+    if (Array.isArray(values)) {
+      for (var i = 0; i < values.length; i++) {
+        addChild(children, values[i])
+      }
     }
-  }
-  else {
-    for (var i = 2; i < arguments.length; i++) {
-      addChild(children, arguments[i])
+    else {
+      for (var i = 2; i < arguments.length; i++) {
+        addChild(children, arguments[i])
+      }
     }
   }