Let's start from this implementation which passes the current tests.

Just to have an idea, here are the benchs:

**Before**
```sh
NANOBENCH version 2
> node

# vnode with no child 1M times
ok ~315 ms (0 s + 315373353 ns)

# vnode with a single child 1M times
ok ~1.35 s (1 s + 353013098 ns)

# vnode with a single array child 1M times
ok ~1.66 s (1 s + 657042054 ns)

all benchmarks completed
ok ~3.33 s (3 s + 325428505 ns)
```

**After**
```sh
NANOBENCH version 2
> node

# vnode with no child 1M times
ok ~287 ms (0 s + 287287753 ns)

# vnode with a single child 1M times
ok ~692 ms (0 s + 692102600 ns)

# vnode with a single array child 1M times
ok ~822 ms (0 s + 822467710 ns)

all benchmarks completed
ok ~1.8 s (1 s + 801858063 ns)
```