@jbucaran 2017-04-15
@ngryman Good start! Remember that when using hyperx, children can be arrays that contain nested arrays too.