@jbucaran 2017-04-15
Here is the implementation that I was able to find that works with both hyperx and jsx and drops the stack:

```js
export default function (tag, data) {
  var children = []

  for (var i = 2; i < arguments.length;) {
    push(arguments[i++], children)
  }

  return typeof tag === "string"
    ? {
      tag: tag,
      data: data || {},
      children: children
    }
    : tag(data, children)

  function push(node) {
    if (Array.isArray(node)) {
      for (var i = 0; i < node.length; i++) {
        push(node[i], children)
      }
    } else if (node != null && node !== true && node !== false) {
      children.push(node)
    }
  }
}
```

It's small, but performance was bad, and I couldn't optimize it. I hope it gives you some ideas.