@ngryman 2017-04-15
@jbucaran Coverage drops because no tests expose the way `hyperx` pass children as 3rd parameter and not as splats.