@ngryman 2017-04-15
New benchs, I can't really do better I think:
```js
NANOBENCH version 2
> node

# vnode with no child 1M times
ok ~277 ms (0 s + 277363055 ns)

# vnode with a single child 1M times
ok ~905 ms (0 s + 905257714 ns)

# vnode with a single array child 1M times
ok ~934 ms (0 s + 934230712 ns)

all benchmarks completed
ok ~2.12 s (2 s + 116851481 ns)
```

But I'm gonna tease you @jbucaran, if we enforce every node to be the exact same object (even for text nodes), here is what we get:
```sh
NANOBENCH version 2
> node

# vnode with no child 1M times
ok ~329 ms (0 s + 328797485 ns)

# vnode with a single child 1M times
ok ~547 ms (0 s + 546626621 ns)

# vnode with a single array child 1M times
ok ~552 ms (0 s + 552139589 ns)

all benchmarks completed
ok ~1.43 s (1 s + 427563695 ns)
```

I could explain further why such a big boost, but basically the current vdom implementation mixes different types of nodes, it can be `string`s or `object`. If we *statically type* the node object by creating a constructor for it and we make sure that the tree only manipulates instances of this constructor, then `v8` and friends can optimize this heavily because it always manipulates data of the same shape. Hence the boost.