@jbucaran 2017-04-16
> @ngryman Coverage drops because no tests expose the way hyperx pass children as 3rd parameter and not as splats.

Then we need to add those tests :)