@jbucaran 2017-04-16
@ngryman Your final implementation is very similar to the recursive one I created, but bigger. Performance is better than that one though! 👍 

However, this version is not faster than the current one on master either.