@jbucaran 2017-04-16
> @ngryman if we enforce every node to be the exact same object (even for text nodes), here is what we get...

Sounds good, but that does not seem to be what this PR is addressing.