@ngryman 2017-04-16
> Then we need to add those tests :)

They have been added, 100% is back :)

> @ngryman Your final implementation is very similar to the recursive one I created, but bigger.

Because it treats the 2 functions signatures in an optimal way:
 - `h(tag, data, children)`
 - `h(tag, data, ...children)`

`arguments` object must be used very carefully if you don't want to kill perfs. I also added some `if`s to avoid to call `Array.isArray` with `null` values which deoptimize the code.

> Now, this version is not really faster than the current one on master (mine wasn't either which is why I didn't merge it).

 - `master` version: `ok ~3.34 s (3 s + 338336241 ns)`
 - `h-optimized` version: `ok ~2.12 s (2 s + 116851481 ns)`

It depends how you bench it. What is your bench I'm curious?

> Sounds good, but that does not seem to be what this PR is addressing no? Sorry, I am terrible at discussing two different things at the same time.

Yes it's just the next step. I also give too much information lately instead of giving you a concise report. It's because it's a complex subject also.