@jbucaran 2017-04-16
@ngryman I use https://github.com/krausest/js-framework-benchmark.