@jbucaran 2017-04-16
> @ngryman Because it treats the 2 functions signatures in an optimal way:

`tag` can be either a String or a Function, no? Doesn't that defeat what V8?