@jbucaran 2017-04-16
Is it necessary to repeat the `for` loop 3 times? What about this implementation that does not change bundle size?

```js
export default function(tag, data, values) {
  var children = []

  if (arguments.length > 2) {
    if (Array.isArray(values)) {
      loop(children, values, 0)
    } else {
      loop(children, arguments, 2)
    }
  }

  return typeof tag === "string"
    ? {
        tag: tag,
        data: data || {},
        children: children
      }
    : tag(data, children)
}

function loop(children, child, i) {
  for (; i < child.length; ) {
    addChild(children, child[i++])
  }
}

function addChild(children, child) {
  if (child != null && child !== true && child !== false) {
    if (Array.isArray(child)) {
      loop(children, child, 0)
    } else {
      children.push(typeof child === "number" ? child + "" : child)
    }
  }
}
```