@ngryman 2017-04-16
@jbucaran Yes, like @leeoniya says, sometimes we have to trade bytes for performance.

Your proposed solution gives this:
```js
```