@ngryman 2017-04-16
I'm not saying that there is not a better implementation than mine though, not at all 😆