@ngryman 2017-04-16
@jbucaran btw, do you test your bundle size gziped? `gzip` logically handles redundancies, multiple for loops should be *grasped* by `gzip` somehow.