@ngryman 2017-04-16
Your version give a gziped bundle of `1702B`, mine `1710B`.