@jbucaran 2017-04-16
Bummer. I just ran it through the benchmark suite and it was definitely slower. 👍 

I'm still not getting any perf gains from yours against the one on master though, so I am not compelled to merge this one, at least not until other priorities are sorted out like improving the patch algorithm.