@ngryman 2017-04-16
> Bummer. I just ran it through the benchmark suite and it was definitely slower. 👍

I'm curious how this version could be slower than the one on `master`, it does not make sense.
Well, let's define what are the priorities then 😆