@jbucaran 2017-04-16
Your version is certainly not slower than the current one on master, but it's not faster either. I'm using js-framework-benchmark. And they both score 1.35.