@leeoniya 2017-04-17
js-framework-benchmark won't always show improvements from micro-optimizations since timings in all its metrics are dominated by dom ops & relayouts/reflows/repaints. that's why it's good to have an additional bench that stresses framework internals without any effect on the dom.