@jbucaran 2017-04-17
> @ngryman Well, let's define what are the priorities then 😆

Which is why I think we can leave this PR open for when the time is right. It's just weird to merge this now, that's what I mean.