@jbucaran 2017-04-17
@developit Can you add your feedback here? 
Your https://github.com/developit/preact/blob/master/src/h.js is very similar and perhaps could benefit from this discussion too.