@ngryman 2017-04-17
To make things short:
 - The PR version is faster but when tested in a global benchmark there is no difference.
 - The true perf gain will come when VNodes will all share the same constructor (hidden class: https://gist.github.com/twokul/9501770).
 - It's the start of several passes of optimizations, imho we have to start somewhere :smile: