@jbucaran 2017-04-17
@ngryman Thanks for the summary and the short article. Just what I needed for my low attention span issues 😄.

And yes, we have to start somewhere, and that somewhere is possibly patch, who's doing all the heavy lifting around here. 💪